+++ logaperture-control-jmx/src/test/java/org/logaperture/control/jmx/LevelControlMXBeanImplTest.java
@@ -54,7 +54,7 @@ class LevelControlMXBeanImplTest {
     }
 
     private static LevelControlMXBeanImpl bean(FakeLevelControlOperations fake) {
-        return new LevelControlMXBeanImpl(fake, fake, fake, fake, fake, fake, fake);
+        return new LevelControlMXBeanImpl(fake, fake, fake, fake, fake, fake, fake, fake);
     }
 
     @Test